+++ database/migrations/2020_01_12_151501_user_messages.php
@@ -14,9 +14,9 @@ class UserMessages extends Migration
     public function up()
     {
          schema::create('user_messages',function(Blueprint $table){
-            $table->increments('id');
-            $table->integer('sender_id')->unsigned();
-            $table->integer('receiver_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('sender_id')->unsigned();
+            $table->bigInteger('receiver_id')->unsigned();
             $table->integer('message_type');
             $table->string('title');
             $table->text('message');
@@ -32,8 +32,8 @@ class UserMessages extends Migration
 
         schema::create('user_message_replies',function(Blueprint $table){
             $table->increments('id');
-            $table->integer('message_id')->unsigned();;
-            $table->integer('sender_id')->unsigned();;
+            $table->bigInteger('message_id')->unsigned();;
+            $table->bigInteger('sender_id')->unsigned();;
             $table->text('message');
             $table->boolean('isSeen');
